Added real_time argument to getMeasure()
By default, for scales other than max, the timestamp of values is
offset by +scale/2. For example, for a scale of 1day the times amps
will be at noon. The real_time causes the timestamp to be the beginning
of the period. Setting realtime to True keeps the timestamp from
potentially exceeding date_end.