Interim release : New dashboard support
Current changes in progress at Netatmo are based on introduction of a data dashboard in the device list request that I expect to be fully supported in the API (congrats to Netatmo).

This is a very welcome addition to replace the previous unsupported last_data_store structure and it even includes some new pieces of information that are of interesting use.

This work is not yet published by netatmo so changes may still occurs. I publish this interim release to provide a way to restore a working API but you have to know that it can break again at any time until Netatmo stabilized the new development.