Removed extra function from setup.py
Removed the install_scripts_and_symlinks function from setup.py. This
appears to be attempting to link executable scripts to somewhere or
some nonsense. Whatever. It totally breaks on *nix systems without
elevated privileges and it isn’t even the correct way to do what it is
trying to do.

Also, the package still seems to work with it removed. So… Gonzo.